requirements.txt
pandas==2.3.3
numpy==2.3.5
pyarrow==22.0.0
requests==2.32.5
pyyaml==6.0.3
pytest==9.0.2
duckdb>=1.1.3

